+++ aftiktuna-rust/src/location.rs
@@ -284,7 +284,7 @@ pub(crate) fn spawn_starting_crew_and_ship(
     );
 
     'add_crew: for profile in crew_iter {
-        while position::check_is_pos_blocked(iter_pos, &world).is_err() {
+        while position::check_is_pos_blocked(None, iter_pos, &world).is_err() {
             let Some(new_pos) = iter_pos.try_offset(1, &world) else {
                 eprintln!(
                     "Tried initializing a crew that is too large. Not all crew members will be added."
@@ -397,7 +397,7 @@ fn deploy_crew_at_new_location(start_pos: Pos, state: &mut GameState) {
     crew_members.swap(0, controlled_index);
     let direction = Direction::towards_center(start_pos, world);
     for character in crew_members {
-        if let Err(blockage) = position::check_is_pos_blocked(start_pos, world) {
+        if let Err(blockage) = position::check_is_pos_blocked(Some(character), start_pos, world) {
             let push_result = blockage.try_push(direction, world);
             if push_result.is_err() {
                 break;